Module todo::todo_errors

Error Handling

In rusty todo there are 3 categories of errors:

Recoverable errors;

Unrecoverable errors that trigger a user notification;

Unrecoverable erros that cause panic.

Recoverable Errors

Possible actions: Propagate (if in not in main) OR Solve error (if in main).

Note: All erors in the rusty_todo lib crate should be propagated as a ToDoError!

Unrecoverable Errors (user notification)

Possible Actions: Notify user, end program without panicking.

Note: This kind of error should be used when the error is likely caused by the user and there is no relevant information to be displayed from a developer standpoint (ex: user used command line arguments incorrectly). Notificating the user will sufice.

Unrecoverable errors (panic)

Possible actions: Notify user and panic OR panic.

Note: This are technical errors that are not likely user triggered and are relevant from a developer standpoint. This should be the most common type of error during development and the least common after release. This errors most likely indicate bugs.

Enums

ToDoError